Adapt the data given by the fuel gauge PCB to the PC
In the script BatteryFunctions.ino the last parameter of the
RaedBatParam has to be a pointer dereference (*) likewise the last
parameter of the functions that call this one, that has to be a pointer
reference (&). That has to be done in this way because in C language
a function can not return an array of data. But, if you work inside the
function with a pointer, that information is stored properly and you can
get it outside the function.

In the script messenger.py the unpacking of soc is done in this way
becauase the PC recieves 4 bytes of data and the first and the third
always are 0x00 and the data is contained in the other two bytes.