Implement the Kalman filter on the multiplecamera module
Prediction and measurement noises were established according to
experimental data.

The time difference between iterations is collected in order to be able
to call the prediction method. This method needs as well the speed
inputs sent to the UGV, which are received through a socket.

The pose and Kalman processing routines are ignored at the beginning,
before the triangle is firstly detected. Moreover, when the UGV is not
detected by the cameras (detected_triangle is False), the measurement
noise is set to a very high value; and when the speeds are not received
through the socket, its noise is set to a high value too.