Redistribute IPs and get all HSV components in renamed rgb2hsv script
Previously, it was only calculated the Hue component of the HSV.
However, the other 2 components (Saturation and Brightness) offer very
useful information for additional image processing algorithms that could
be implemented in the short term.

The new redistribution of the IP folder basically consists on moving the
strictly image processing algorithms to a new sub-folder, instead of
having them embedded on the camera_controller sub-folder.