Add 2 new image processing hardware components
The first, image_processing, is the top level component of the package,
and it integrates the rest of the components inside the package, in
order to implement the desired image pipeline (currently, RGB to HSV
conversion, followed by a binarization).

The second, hsv2bin, gets image pixels with HSV format and converts them
to binary values, being '1' if the pixel has certain user-defined
hue, saturation and brightness features.